pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = ""
authors = ["Shuja"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
djoser = "^2.2.3"
timedelta = "^2020.12.3"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"
watchdog = "^5.0.3"
pillow = "^11.0.0"
stripe = "^11.1.1"
dj-database-url = "^2.3.0"
black = "^24.10.0"
django-cors-headers = "^4.6.0"
cloudinary = "^1.41.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"